Move files, read values from context instead of props
